+++ wicket/components-wicket/src/main/java/nl/rotterdam/design_system/wicket/components/alert/utrecht/UtrechtAlert.java
@@ -1,15 +1,17 @@
 package nl.rotterdam.design_system.wicket.components.alert.utrecht;
 
+import nl.rotterdam.design_system.wicket.components.component_state.Community;
 import nl.rotterdam.design_system.wicket.components.component_state.NlComponentState;
 import org.apache.wicket.markup.html.basic.Label;
 import org.apache.wicket.markup.html.panel.Panel;
 import org.apache.wicket.model.IModel;
 
+import static nl.rotterdam.design_system.wicket.components.component_state.Community.UTRECHT;
 import static nl.rotterdam.design_system.wicket.components.component_state.EstafetteState.COMMUNITY;
 import static nl.rotterdam.design_system.wicket.components.component_state.WicketState.UNSTABLE;
 
 
-@NlComponentState(wicketState = UNSTABLE, estafetteState = COMMUNITY)
+@NlComponentState(wicketState = UNSTABLE, estafetteState = COMMUNITY, htmlCssImplementedBy = UTRECHT)
 public class UtrechtAlert extends Panel {
 
     public static final String SLOT_ID = "slot";
