+++ wicket/components-wicket/src/main/java/nl/rotterdam/design_system/wicket/components/column_layout/utrecht/UtrechtColumnLayoutBorder.java
@@ -3,11 +3,12 @@ package nl.rotterdam.design_system.wicket.components.column_layout.utrecht;
 import nl.rotterdam.design_system.wicket.components.component_state.NlComponentState;
 import org.apache.wicket.markup.html.border.Border;
 
+import static nl.rotterdam.design_system.wicket.components.component_state.Community.UTRECHT;
 import static nl.rotterdam.design_system.wicket.components.component_state.EstafetteState.COMMUNITY;
 import static nl.rotterdam.design_system.wicket.components.component_state.WicketState.NEEDS_REFACTORING;
 
 
-@NlComponentState(wicketState = NEEDS_REFACTORING, estafetteState = COMMUNITY)
+@NlComponentState(wicketState = NEEDS_REFACTORING, estafetteState = COMMUNITY, htmlCssImplementedBy = UTRECHT)
 public class UtrechtColumnLayoutBorder extends Border {
 
     public UtrechtColumnLayoutBorder(String id) {
