+++ wicket/components-wicket/src/main/java/nl/rotterdam/design_system/wicket/components/component_state/Community.java
@@ -1,0 +1,19 @@
+package nl.rotterdam.design_system.wicket.components.component_state;
+
+public enum Community {
+    AMSTERDAM("Gemeente Amsterdam"),
+    DEN_HAAG("Gemeente Den Haag"),
+    EINDHOVEN("Gemeente Eindhoven"),
+
+    KERNTEAM("Kernteam NL Design System"),
+    LUX("Logius"),
+    ROTTERDAM("Gemeente Rotterdam"),
+    RVO("Rijksdienst voor Ondernemend Nederland"),
+    TILBURG("Gemeente Tilburg"),
+    UTRECHT("Gemeente Utrecht");
+
+
+    Community(String name) {
+
+    }
+}
