+++ wicket/components-wicket/src/main/java/nl/rotterdam/design_system/wicket/components/component_state/NlComponentState.java
@@ -4,7 +4,11 @@ package nl.rotterdam.design_system.wicket.components.component_state;
  * Annotates the state of a component in both Wicket and NL Design Estafette Model state terms.
  */
 public @interface NlComponentState {
+
     WicketState wicketState();
     EstafetteState estafetteState();
+
+    Community htmlCssImplementedBy();
+
 
 }
