+++ wicket/components-wicket/src/main/java/nl/rotterdam/design_system/wicket/components/form_field/utrecht/UtrechtFormFieldCss.java
@@ -1,8 +1,14 @@
 package nl.rotterdam.design_system.wicket.components.form_field.utrecht;
 
+import nl.rotterdam.design_system.wicket.components.component_state.Community;
 import nl.rotterdam.design_system.wicket.components.css_class_names.CssClassNames;
 import org.jspecify.annotations.NonNull;
 
+
+/**
+ * Intentionally did not rename to NL, kept Utrecht in it. It has public visibility, can be used by library users,
+ * and should only be used for Form Fields implemented by {@link Community} Utrecht.
+ */
 public enum UtrechtFormFieldCss implements CssClassNames {
 
     DEFAULT("utrecht-form-field"),
