+++ wicket/components-wicket/src/main/java/nl/rotterdam/design_system/wicket/components/heading/utrecht/UtrechtHeading.java
@@ -7,10 +7,11 @@ import org.apache.wicket.markup.html.basic.Label;
 import org.apache.wicket.model.IModel;
 import org.apache.wicket.model.Model;
 
+import static nl.rotterdam.design_system.wicket.components.component_state.Community.UTRECHT;
 import static nl.rotterdam.design_system.wicket.components.component_state.EstafetteState.COMMUNITY;
 import static nl.rotterdam.design_system.wicket.components.component_state.WicketState.UNSTABLE;
 
-@NlComponentState(wicketState = UNSTABLE, estafetteState = COMMUNITY)
+@NlComponentState(wicketState = UNSTABLE, estafetteState = COMMUNITY, htmlCssImplementedBy = UTRECHT)
 public class UtrechtHeading extends Label {
 
     public UtrechtHeading(String id, int level) {
