+++ wicket/components-wicket/src/main/java/nl/rotterdam/design_system/wicket/components/icon/utrecht/UtrechtIconBorder.java
@@ -3,11 +3,12 @@ package nl.rotterdam.design_system.wicket.components.icon.utrecht;
 import nl.rotterdam.design_system.wicket.components.component_state.NlComponentState;
 import org.apache.wicket.markup.html.border.Border;
 
+import static nl.rotterdam.design_system.wicket.components.component_state.Community.UTRECHT;
 import static nl.rotterdam.design_system.wicket.components.component_state.EstafetteState.COMMUNITY;
 import static nl.rotterdam.design_system.wicket.components.component_state.WicketState.NEEDS_REFACTORING;
 
 
-@NlComponentState(wicketState = NEEDS_REFACTORING, estafetteState = COMMUNITY)
+@NlComponentState(wicketState = NEEDS_REFACTORING, estafetteState = COMMUNITY, htmlCssImplementedBy = UTRECHT)
 public class UtrechtIconBorder extends Border {
 
     public UtrechtIconBorder(String id) {
