+++ wicket/components-wicket/src/main/java/nl/rotterdam/design_system/wicket/components/link/utrecht/UtrechtLink.java
@@ -6,10 +6,11 @@ import org.apache.wicket.markup.html.link.Link;
 import org.apache.wicket.model.IModel;
 import org.apache.wicket.request.component.IRequestablePage;
 
+import static nl.rotterdam.design_system.wicket.components.component_state.Community.UTRECHT;
 import static nl.rotterdam.design_system.wicket.components.component_state.EstafetteState.COMMUNITY;
 import static nl.rotterdam.design_system.wicket.components.component_state.WicketState.NEEDS_REFACTORING;
 
-@NlComponentState(wicketState = NEEDS_REFACTORING, estafetteState = COMMUNITY)
+@NlComponentState(wicketState = NEEDS_REFACTORING, estafetteState = COMMUNITY, htmlCssImplementedBy = UTRECHT)
 public class UtrechtLink<T, C extends IRequestablePage> extends Link<T> {
 
     public Class<C> webPage = null;
