+++ wicket/components-wicket/src/main/java/nl/rotterdam/design_system/wicket/components/link/utrecht/UtrechtLinkBehavior.java
@@ -1,6 +1,7 @@
 package nl.rotterdam.design_system.wicket.components.link.utrecht;
 
 import static nl.rotterdam.design_system.wicket.components.CssReferenceHeaderItems.cssReferenceHeaderItem;
+import static nl.rotterdam.design_system.wicket.components.component_state.Community.UTRECHT;
 import static nl.rotterdam.design_system.wicket.components.component_state.EstafetteState.COMMUNITY;
 import static nl.rotterdam.design_system.wicket.components.component_state.WicketState.NEEDS_REFACTORING;
 
@@ -10,7 +11,7 @@ import org.apache.wicket.markup.head.CssReferenceHeaderItem;
 import org.apache.wicket.markup.html.link.AbstractLink;
 
 
-@NlComponentState(wicketState = NEEDS_REFACTORING, estafetteState = COMMUNITY)
+@NlComponentState(wicketState = NEEDS_REFACTORING, estafetteState = COMMUNITY, htmlCssImplementedBy = UTRECHT)
 public class UtrechtLinkBehavior extends TagNameClassComponentBehavior<AbstractLink> {
 
     public static final CssReferenceHeaderItem CSS = cssReferenceHeaderItem(
