+++ wicket/components-wicket/src/main/java/nl/rotterdam/design_system/wicket/components/ordered_list/utrecht/UtrechtOrderedList.java
@@ -3,10 +3,11 @@ package nl.rotterdam.design_system.wicket.components.ordered_list.utrecht;
 import nl.rotterdam.design_system.wicket.components.component_state.NlComponentState;
 import org.apache.wicket.markup.html.panel.Panel;
 
+import static nl.rotterdam.design_system.wicket.components.component_state.Community.UTRECHT;
 import static nl.rotterdam.design_system.wicket.components.component_state.EstafetteState.COMMUNITY;
 import static nl.rotterdam.design_system.wicket.components.component_state.WicketState.NEEDS_REFACTORING;
 
-@NlComponentState(wicketState = NEEDS_REFACTORING, estafetteState = COMMUNITY)
+@NlComponentState(wicketState = NEEDS_REFACTORING, estafetteState = COMMUNITY, htmlCssImplementedBy = UTRECHT)
 public class UtrechtOrderedList extends Panel {
 
     public UtrechtOrderedList(String id) {
