+++ wicket/components-wicket/src/main/java/nl/rotterdam/design_system/wicket/components/preserve_data/utrecht/UtrechtPreserveDataLabel.java
@@ -6,6 +6,7 @@ import org.apache.wicket.markup.ComponentTag;
 import org.apache.wicket.markup.html.basic.Label;
 import org.apache.wicket.model.IModel;
 
+import static nl.rotterdam.design_system.wicket.components.component_state.Community.UTRECHT;
 import static nl.rotterdam.design_system.wicket.components.component_state.EstafetteState.COMMUNITY;
 import static nl.rotterdam.design_system.wicket.components.component_state.WicketState.NEEDS_REFACTORING;
 import static nl.rotterdam.design_system.wicket.components.preserve_data.utrecht.UtrechtPreserveDataCss.DEFAULT;
@@ -16,7 +17,7 @@ import static nl.rotterdam.design_system.wicket.components.preserve_data.utrecht
  *  <bdi translate="no" class="utrecht-preserve-data" wicket:id="name">Van Bergenhenegouwen</bdi>
  */
 
-@NlComponentState(wicketState = NEEDS_REFACTORING, estafetteState = COMMUNITY)
+@NlComponentState(wicketState = NEEDS_REFACTORING, estafetteState = COMMUNITY, htmlCssImplementedBy = UTRECHT)
 public class UtrechtPreserveDataLabel extends Label {
 
     public UtrechtPreserveDataLabel(String id, IModel<?> model) {
